# PERF: Optimize TemplateModelLexicalParser.cs